new news things

+++ src/main/java/com/andromeda/araserver/util/NewsData.kt
@@ -1,2 +1,2 @@
 package com.andromeda.araserver.util
-data class NewsData(val title:String, val info:String, val smallInfo:String, val link:String, val topics:ArrayList<String>, val date:Int)
+data class NewsData(val title:String, val info:String, val smallInfo:String, val link:String, val topics:ArrayList<String>, val date:Long)
